cifsd: move find_fp_using_inode()
find_fp_using_inode() belongs to fh.c and fh.h. Move it out
of misc.c and glob.h (which must die).
